requirements.txt
flask==3.0.0
chromadb==0.4.22
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10
pypdf==3.17.4
markdown==3.5.1
python-dotenv==1.0.0
requests==2.31.0
huggingface-hub==0.19.4
